+++ src/utils/Acc.js
@@ -3,8 +3,7 @@
  * @param {string} hash 
  */
 export function getImageUrl(hash) {
-    const url = API.substring(0, API.length - 6);
-    return `${url}Service/WxGetFile.ashx?hash=${hash}`;
+    return `${EAP}Service/WxGetFile.ashx?hash=${hash}`;
 }
 
 /**
@@ -12,8 +11,7 @@ export function getImageUrl(hash) {
  * @param {string} hash 
  */
 export function getThumbUrl(hash) {
-    const url = API.substring(0, API.length - 6);
-    return `${url}Service/WxGetFile.ashx?hash=${hash}&imgSize=small`;
+    return `${EAP}Service/WxGetFile.ashx?hash=${hash}&imgSize=small`;
 }
 
 /**
@@ -21,8 +19,7 @@ export function getThumbUrl(hash) {
  * @param {int} accId 
  */
 export function getPreviewPath(accId) {
-    const url = API.substring(0, API.length - 6);
-    return `${url}Base/Lib/WxFilePreview.aspx?AccId=${accId}&isMobile=1`;
+    return `${EAP}Base/Lib/WxFilePreview.aspx?AccId=${accId}&isMobile=1`;
 }
 /**
  * 根据hash获取图片地址
@@ -63,7 +60,7 @@ export function getImageFromContent(oUrl) {
     // 如果是Base64图片
     if(oUrl.indexOf('../../') < 0) return oUrl;
     // WxAccessoryDisplay
-    if(oUrl.indexOf('WxAccessoryDisplay') > 0) return `${API.substring(0, API.length - 6)}Common/WxAccessoryDisplay.aspx${oUrl.substring(oUrl.indexOf('?'))}`;
+    if(oUrl.indexOf('WxAccessoryDisplay') > 0) return `${EAP}Common/WxAccessoryDisplay.aspx${oUrl.substring(oUrl.indexOf('?'))}`;
     // 其他情况(大多数)
-    return `${API.substring(0, API.length - 6)}Common/WxGetFile.aspx${oUrl.substring(oUrl.indexOf('?'))}`;
+    return `${EAP}Common/WxGetFile.aspx${oUrl.substring(oUrl.indexOf('?'))}`;
 }