+++ src/utils/backToHome.js
@@ -1,3 +1,8 @@
+import { DEV_MODE } from '../constant';
 export default function backToHome() {
-    location.href = `${API.slice(0, API.indexOf('wxapi'))}Base/Main/WxTouchMain.aspx`;
+    if(DEV_MODE) {
+        const config = require('../../../../config/default.json');
+        return location.href = `${config.origin}/Base/Main/WxTouchMain.aspx`;
+    }
+    return location.href = `${EAP}Base/Main/WxTouchMain.aspx`;
 }